Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -3,7 +3,7 @@
 int main() {
     int radius;
     float area;
-    int square;
+    float square;
     scanf("%d",&radius);
     square= radius ** 2;
     area=3.14*square;
